+++ typescript/registry/RELEASES.md
@@ -17,3 +17,13 @@ Based on:
 - [typescript v0.3.4] typescript/registry
 ### Releases
 - [NPM v0.3.4] https://www.npmjs.com/package/@smithery/registry/v/0.3.4 - typescript/registry
+
+## 2025-05-11 06:05:38
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.543.4 (2.598.22) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v0.3.5] typescript/registry
+### Releases
+- [NPM v0.3.5] https://www.npmjs.com/package/@smithery/registry/v/0.3.5 - typescript/registry